[Updating] Create a dedicated executable project for updating procedures
