+++ src/Update/Resources.resx
@@ -1,0 +1,91 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<root>
+  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
+    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
+    <xsd:element name="root" msdata:IsDataSet="true">
+      <xsd:complexType>
+        <xsd:choice maxOccurs="unbounded">
+          <xsd:element name="metadata">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" />
+              </xsd:sequence>
+              <xsd:attribute name="name" use="required" type="xsd:string" />
+              <xsd:attribute name="type" type="xsd:string" />
+              <xsd:attribute name="mimetype" type="xsd:string" />
+              <xsd:attribute ref="xml:space" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="assembly">
+            <xsd:complexType>
+              <xsd:attribute name="alias" type="xsd:string" />
+              <xsd:attribute name="name" type="xsd:string" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="data">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
+              </xsd:sequence>
+              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
+              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
+              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
+              <xsd:attribute ref="xml:space" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="resheader">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+              </xsd:sequence>
+              <xsd:attribute name="name" type="xsd:string" use="required" />
+            </xsd:complexType>
+          </xsd:element>
+        </xsd:choice>
+      </xsd:complexType>
+    </xsd:element>
+  </xsd:schema>
+  <resheader name="resmimetype">
+    <value>text/microsoft-resx</value>
+  </resheader>
+  <resheader name="version">
+    <value>2.0</value>
+  </resheader>
+  <resheader name="reader">
+    <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </resheader>
+  <resheader name="writer">
+    <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </resheader>
+  <data name="DOTNET_CORE_DOWNLOAD_FAILURE" xml:space="preserve">
+    <value>Couldn't download .NET Core Desktop Runtime 3.1, please install it manually.</value>
+  </data>
+  <data name="DOTNET_CORE_DOWNLOAD_FAILURE_TITLE" xml:space="preserve">
+    <value>PowerToys installation error</value>
+  </data>
+  <data name="GITHUB_NEW_VERSION_AVAILABLE" xml:space="preserve">
+    <value>An update to PowerToys is available.</value>
+  </data>
+  <data name="GITHUB_NEW_VERSION_UPDATE_NOW" xml:space="preserve">
+    <value>Update now</value>
+  </data>
+  <data name="UNINSTALLATION_UNKNOWN_ERROR" xml:space="preserve">
+    <value>Error: please uninstall the previous version of PowerToys manually.</value>
+  </data>
+  <data name="GITHUB_NEW_VERSION_AVAILABLE_OFFER_VISIT" xml:space="preserve">
+    <value>An update to PowerToys is available. Visit our GitHub page to update.</value>
+  </data>
+  <data name="GITHUB_NEW_VERSION_MORE_INFO" xml:space="preserve">
+    <value>More info...</value>
+  </data>
+  <data name="TOAST_TITLE" xml:space="preserve">
+    <value>PowerToys Update</value>
+  </data>
+  <data name="OFFER_UNINSTALL_MSI" xml:space="preserve">
+    <value>We've detected a previous installation of PowerToys. Would you like to remove it?</value>
+  </data>
+  <data name="OFFER_UNINSTALL_MSI_TITLE" xml:space="preserve">
+    <value>PowerToys: uninstall previous version?</value>
+  </data>
+</root>
