+++ src/runner/action_runner_utils.h
@@ -16,7 +16,6 @@ namespace cmdArg
     const inline wchar_t* UPDATE_STAGE2_RESTART_PT = L"restart";
     const inline wchar_t* UPDATE_STAGE2_DONT_START_PT = L"dont_start";
 
-    const inline wchar_t* UNINSTALL_MSI = L"-uninstall_msi";
     const inline wchar_t* RUN_NONELEVATED = L"-run-non-elevated";
 
     const inline wchar_t* UPDATE_REPORT_SUCCESS = L"-report_update_success";
